Split rule
