+++ accessibility-checker-engine/help-v4/en-US/elem_lang_valid.html
@@ -74,9 +74,7 @@ For example:
 
 ### About this requirement
 
-* [IBM 3.1.1 Language of page](https://www.ibm.com/able/requirements/requirements/#3_1_1)
 * [IBM 3.1.2 Language of parts](https://www.ibm.com/able/requirements/requirements/#3_1_2)
-* [WCAG technique H57](https://www.w3.org/WAI/WCAG21/Techniques/html/H57)
 * [WCAG technique H58](https://www.w3.org/WAI/WCAG21/Techniques/html/H58)
 
 ### Who does this affect?
