+++ accessibility-checker-engine/src/v4/rules/index.ts
@@ -28,6 +28,7 @@ export * from "./combobox_popup_reference"
 export * from "./combobox_version"
 export * from "./debug_paths"
 export * from "./detector_tabbable"
+export * from "./elem_lang_valid"
 export * from "./element_attribute_deprecated"
 export * from "./element_tabbable_role_valid"
 export * from "./element_tabbable_visible"
@@ -142,7 +143,6 @@ export * from "./WCAG20_Blink_AlwaysTrigger"
 export * from "./WCAG20_Body_FirstAContainsSkipText_Native_Host_Sematics"
 export * from "./WCAG20_Body_FirstASkips_Native_Host_Sematics"
 export * from "./WCAG20_Doc_HasTitle"
-export * from "./WCAG20_Elem_Lang_Valid"
 export * from "./WCAG20_Elem_UniqueAccessKey"
 export * from "./WCAG20_Embed_HasNoEmbed"
 export * from "./WCAG20_Fieldset_HasLegend"
