+++ accessibility-checker-engine/test/v2/checker/accessibility/rules/WCAG20_Elem_Lang_Valid_ruleunit/ACT_bf051a_e806126ba40cc2ee3da9d01afb217bded3b816bc.html
@@ -12,10 +12,10 @@
     </main>
     <script>
       UnitTest = {
-        ruleIds: ["WCAG20_Elem_Lang_Valid"],
+        ruleIds: ["html_lang_valid"],
         results: [
           {
-            "ruleId": "WCAG20_Elem_Lang_Valid",
+            "ruleId": "html_lang_valid",
             "value": [
               "INFORMATION",
               "FAIL"
