+++ accessibility-checker-engine/test/v2/checker/accessibility/rules/WCAG20_Elem_Lang_Valid_ruleunit/ACT_de46e4_4c655a5da71710e6ca6245186af7f5963e89b332.html
@@ -16,26 +16,10 @@
     </main>
     <script>
       UnitTest = {
-        ruleIds: ["WCAG20_Elem_Lang_Valid"],
+        ruleIds: ["elem_lang_valid"],
         results: [
           {
-            "ruleId": "WCAG20_Elem_Lang_Valid",
-            "value": [
-              "INFORMATION",
-              "FAIL"
-            ],
-            "path": {
-              "dom": "/html[1]",
-              "aria": "/document[1]"
-            },
-            "reasonId": "Fail_1",
-            "message": "Specified 'lang' attribute does not include a valid primary language",
-            "messageArgs": [],
-            "apiArgs": [],
-            "category": "Accessibility"
-          },
-          {
-            "ruleId": "WCAG20_Elem_Lang_Valid",
+            "ruleId": "elem_lang_valid",
             "value": [
               "INFORMATION",
               "FAIL"
