+++ accessibility-checker-engine/test/v2/checker/accessibility/rules/WCAG20_Html_HasLang_ruleunit/Html-invalidEmptyXmlLang.html
@@ -48,10 +48,10 @@
 
     <script type="text/javascript">
         UnitTest = {
-            ruleIds: ["WCAG20_Html_HasLang", "WCAG20_Elem_Lang_Valid"],
+            ruleIds: ["WCAG20_Html_HasLang", "html_lang_valid"],
             results: [
                 {
-                    "ruleId": "WCAG20_Elem_Lang_Valid",
+                    "ruleId": "html_lang_valid",
                     "value": [
                         "INFORMATION",
                         "FAIL"
@@ -61,7 +61,7 @@
                         "aria": "/document[1]"
                     },
                     "reasonId": "Fail_3",
-                    "message": "Specified 'lang' attribute does not include a valid primary language",
+                    "message": "Specified 'xml:lang' attribute does not include a valid primary language",
                     "messageArgs": [],
                     "apiArgs": [],
                     "category": "Accessibility"
